fix: do not show some images on very fast connections on parallax
Previously, the following sequence of events was possible:

- The browser loaded the HTML document and some parallax images, showing
  them as they were loaded.
- The document finished loading, and the island JavaScript code was run.
- The JavaScript code hide the parallax images again, only to unhide
  them shortly after because they were already loaded.

In other words, a partial parallax effect could show if the images were
loaded before the full document was retrieved, and the component
JavaScript started execution.

Fix that by hiding the images by default, adding a <noscript> fallback
for user agents that do not wish to run JavaScript.

Let's also remove the debugging code that helped catching this issue.